+++ lib/adapters/objection-adapter.ts
@@ -1,6 +1,7 @@
 import {DefaultAdapter} from "./default-adapter";
 
 /* eslint-disable class-methods-use-this */
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 export class ObjectionAdapter extends DefaultAdapter {
 	async save<T>(instance: T, Model: any): Promise<T> {
 		return Model.query().insertAndFetch(instance);