+++ lib/adapters/sequelize-adapter.ts
@@ -1,6 +1,7 @@
 import {DefaultAdapter} from "./default-adapter";
 
 /* eslint-disable class-methods-use-this */
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 export class SequelizeAdapter extends DefaultAdapter {
 	build(Model: any): any {
 		return Model.build();