+++ lib/attributes/relation-attribute.ts
@@ -2,18 +2,22 @@ import {Attribute} from "./attribute";
 import {Evaluator} from "../evaluator";
 
 export class RelationAttribute extends Attribute {
-	fixture: string | string[];
+	fixture: string[];
 	overrides: any[];
 
 	constructor(name: string, fixture: string | string[], overrides: any[]) {
 		super(name, false);
-		this.fixture = fixture;
+		if (Array.isArray(fixture)) {
+			this.fixture = fixture;
+		} else {
+			this.fixture = [fixture];
+		}
 		this.overrides = overrides;
 		this.isRelation = true;
 	}
 
 	evaluate(evaluator: Evaluator): () => Promise<Record<string, any>> {
-		const traitsAndOverrides = [this.fixture, this.overrides].flat();
+		const traitsAndOverrides = Array.prototype.concat(this.fixture, this.overrides);
 		const fixtureName = traitsAndOverrides.shift();
 
 		return async() => {