+++ lib/fixture-riveter.ts
@@ -22,6 +22,8 @@ import {
 	Overrides,
 } from "./types";
 
+import {DefaultStrategyMethods} from "./strategies/default-strategy-methods";
+
 import {
 	isPlainObject,
 	isString,
@@ -50,7 +52,7 @@ export function nameGuard<T>(fixtureName: FixtureName<T>): string {
 	throw new Error(`${fixtureName} isn't the right shape`);
 }
 
-export class FixtureRiveter {
+export class FixtureRiveter implements DefaultStrategyMethods {
 	fixtures: Record<string, Fixture<any>>;
 	traits: Record<string, Trait<any>>;
 	instances: [string, any][];
@@ -257,7 +259,7 @@ export class FixtureRiveter {
 		}
 		const strategy = new StrategyConstructor(sName, this, adapter);
 
-		const assembler = await fixture.run(strategy, overrides);
+		const assembler = await fixture.prepare(strategy, overrides);
 		const instance = await strategy.result<T>(assembler, fixture.model);
 		this.instances.push([name, instance]);
 		return instance;
@@ -299,7 +301,7 @@ export class FixtureRiveter {
 
 	registerStrategy(...rest: any[]): void {
 		let name: string | FixtureName<Strategy>;
-		let strategyClass: Strategy;
+		let strategyClass: typeof Strategy;
 
 		if (isString(rest[0])) {
 			name = rest.shift();
@@ -348,127 +350,16 @@ export class FixtureRiveter {
 
 	// Typescript sucks for dynamically defined methods lol
 	// All of these will be overwritten on instantiation
-	async attributesFor<T = Pojo>(
-		name: FixtureName<T>,
-		traits?: string[],
-		overrides?: Overrides<T>,
-	): Promise<T>;
-
-	async attributesFor<T = Pojo>(
-		name: FixtureName<T>,
-		traitOrOverrides?: string[]|Overrides<T>,
-	): Promise<T>;
-
-	async attributesFor<T>(..._args: any[]): Promise<T> { return undefined as any; }
-
-	async attributesForList<T = Pojo>(
-		name: FixtureName<T>,
-		count: number,
-		traits?: string[],
-		overrides?: Overrides<T>,
-	): Promise<T[]>;
-
-	async attributesForList<T = Pojo>(
-		name: FixtureName<T>,
-		count: number,
-		traitOrOverrides?: string[]|Overrides<T>,
-	): Promise<T[]>;
-
-	async attributesForList<T>(..._args: any[]): Promise<T[]> { return undefined as any; }
-
-	async attributesForPair<T = Pojo>(
-		name: FixtureName<T>,
-		traits?: string[],
-		overrides?: Overrides<T>,
-	): Promise<[T, T]>;
-
-	async attributesForPair<T = Pojo>(
-		name: FixtureName<T>,
-		traitOrOverrides?: string[]|Overrides<T>,
-	): Promise<[T, T]>;
-
-	async attributesForPair<T>(..._args: any[]): Promise<[T, T]> { return undefined as any; }
-
-	async build<T = Pojo>(
-		name: FixtureName<T>,
-		traitOrOverrides?: string[]|Overrides<T>,
-	): Promise<T>;
-
-	async build<T = Pojo>(
-		name: FixtureName<T>,
-		traits?: string[],
-		overrides?: Overrides<T>,
-	): Promise<T>;
+	async attributesFor<T = Pojo>(..._args: any[]): Promise<T> { return undefined as any; }
+	async attributesForList<T = Pojo>(..._args: any[]): Promise<T[]> { return undefined as any; }
+	async attributesForPair<T = Pojo>(..._args: any[]): Promise<[T, T]> { return undefined as any; }
 
 	async build<T = Pojo>(..._args: any[]): Promise<T> { return undefined as any; }
-
-	async buildList<T = Pojo>(
-		name: FixtureName<T>,
-		count: number,
-		traits?: string[],
-		overrides?: Overrides<T>,
-	): Promise<T[]>;
-
-	async buildList<T = Pojo>(
-		name: FixtureName<T>,
-		count: number,
-		traitOrOverrides?: string[]|Overrides<T>,
-	): Promise<T[]>;
-
 	async buildList<T = Pojo>(..._args: any[]): Promise<T[]> { return undefined as any; }
-
-	async buildPair<T = Pojo>(
-		name: FixtureName<T>,
-		traits?: string[],
-		overrides?: Overrides<T>,
-	): Promise<[T, T]>;
-
-	async buildPair<T = Pojo>(
-		name: FixtureName<T>,
-		traitOrOverrides?: string[]|Overrides<T>,
-	): Promise<[T, T]>;
-
 	async buildPair<T = Pojo>(..._args: any[]): Promise<[T, T]> { return undefined as any; }
 
-	async create<T = Pojo>(
-		name: FixtureName<T>,
-		traits?: string[],
-		overrides?: Overrides<T>,
-	): Promise<T>;
-
-	async create<T = Pojo>(
-		name: FixtureName<T>,
-		traitOrOverrides?: string[]|Overrides<T>,
-	): Promise<T>;
-
 	async create<T = Pojo>(..._args: any[]): Promise<T> { return undefined as any; }
-
-	async createList<T = Pojo>(
-		name: FixtureName<T>,
-		count: number,
-		traits?: string[],
-		overrides?: Overrides<T>,
-	): Promise<T[]>;
-
-	async createList<T = Pojo>(
-		name: FixtureName<T>,
-		count: number,
-		traitOrOverrides?: string[]|Overrides<T>,
-	): Promise<T[]>;
-
 	async createList<T = Pojo>(..._args: any[]): Promise<T[]> { return undefined as any; }
-
-	async createPair<T = Pojo>(
-		name: FixtureName<T>,
-		traits?: string[],
-		overrides?: Overrides<T>,
-	): Promise<[T, T]>;
-
-	async createPair<T = Pojo>(
-		name: FixtureName<T>,
-		traitOrOverrides?: string[]|Overrides<T>,
-	): Promise<[T, T]>;
-
 	async createPair<T = Pojo>(..._args: any[]): Promise<[T, T]> { return undefined as any; }
 	/* eslint-enable */
 }