+++ lib/method-missing.ts
@@ -1,14 +1,15 @@
 // This is loosely based on https://github.com/ramadis/unmiss, but that library didn't
 // work so here we are.
 
+// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export function addMethodMissing(definitionProxy: any): any {
 	const handler = {
 		get(target: any, prop: string, receiver: any) {
-			if (Reflect.has(target, prop) || prop === "_methodMissing") {
+			if (Reflect.has(target, prop) || prop === "methodMissing") {
 				return Reflect.get(target, prop, receiver);
 			}
 			return function(...args: any[]) {
-				return Reflect.get(target, "_methodMissing").call(target, prop, ...args);
+				return Reflect.get(target, "methodMissing").call(target, prop, ...args);
 			};
 		},
 	};